Update FxLabelAssignable.java
Rename to avoid name collision with new non-FX general interface designed to work with enums. This one is about enforcing a contract for a class to provide a specific JavaFX observable value for labels, so it's a different sort of use case altogether and should be explicitly named for JavaFX calling contexts.