Reduce boilerplate code and inconsistencies by making a BooleanBinding factory
This code is always identical for every observable model class that adds a BooleanBinding dirty flag, except for the list of observables to use as dependencies, which is more concisely and consistently provided as varargs to this new factory method.

Call the makeBooleanBinding() method in your constructor after making the observable properties that you pass to it. This can always be a curated list as not all observable properties should set a dirty flag. For instance, an ID field.