Start building out a generalized handling framework for mouse gestures and modes
This is just the start of an effort to generalize the framework for gesture and mouse event handling across mouse modes, to avoid copy/paste for boilerplate in new apps. This will take a few revs as the ultimate version will depend on interfaces so that apps can override for specifics and let the toolkit handle the more general aspects.